+++ cmp-worker-web/src/commonMain/kotlin/io/github/mobilebytelabs/worker/web/WebWorkManagerInit.kt
@@ -49,6 +49,7 @@ public fun webWorkManagerFactory(): WorkManagerFactory = WorkManagerFactory { co
             persistenceDbName = config.webConfig.persistenceDbName,
             enableBackgroundSync = config.webConfig.enableBackgroundSync,
             serviceWorkerScript = config.webConfig.serviceWorkerScript,
+            enablePeriodicBackgroundSync = config.webConfig.enablePeriodicBackgroundSync,
         ),
     )
 }